technical-doc-assistant/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
python-dotenv
python-multipart
GitPython
pinecone
langchain
openai